+++ src/main/java/net/additionz/access/TradeOfferAccess.java
@@ -1,0 +1,8 @@
+package net.additionz.access;
+
+public interface TradeOfferAccess {
+
+    void setPlayerExperience(int playerExperience);
+
+    int getPlayerExperience();
+}