+++ src/main/java/net/additionz/config/AdditionConfig.java
@@ -97,6 +97,7 @@ public class AdditionConfig implements ConfigData {
     public boolean ladder_climb_speeding = true;
     @ConfigEntry.Gui.RequiresRestart
     public boolean trinket_slot_arrangement = false;
+    public boolean remove_vanilla_villager_trades = false;
 
     @Comment("Mobs like animals get spawned with a higher chance in a group")
     public boolean passive_entity_group_spawn = true;